Rename vector_utils.py to vector.py
